Improve the \linkitem macro.
It's better we don't consider any argument, and just expand it to \item \url.
By doing this, we can offload the character redefinition to url.
If we take the url as an parameter, it would get interpreted and expanded as is,
which explodes when we have special characters like _ or %.